+++ packages/babel-compiler/package.js
@@ -1,28 +1,24 @@
 Package.describe({
-  name: "babel-compiler",
-  summary: "Parser/transpiler for ECMAScript 2015+ syntax",
+  name: 'babel-compiler',
+  summary: 'Parser/transpiler for ECMAScript 2015+ syntax',
   // Tracks the npm version below.  Use wrap numbers to increment
   // without incrementing the npm version.  Hmm-- Apparently this
   // isn't possible because you can't publish a non-recommended
   // release with package versions that don't have a pre-release
   // identifier at the end (eg, -dev)
-  version: '7.7.0'
+  version: '7.8.0',
 });
 
 Npm.depends({
-  '@meteorjs/babel': '7.13.0',
-  'json5': '2.1.1'
+  '@meteorjs/babel': '7.14.0',
+  json5: '2.1.1',
 });
 
-Package.onUse(function (api) {
+Package.onUse(function(api) {
   api.use('ecmascript-runtime', 'server');
   api.use('modern-browsers');
 
-  api.addFiles([
-    'babel.js',
-    'babel-compiler.js',
-    'versions.js',
-  ], 'server');
+  api.addFiles(['babel.js', 'babel-compiler.js', 'versions.js'], 'server');
 
   api.export('Babel', 'server');
   api.export('BabelCompiler', 'server');